BOCM-358 - moves state machine test files to own folder
